--- a/pkg/loadbalancer/loadbalancer_test.go
+++ b/pkg/loadbalancer/loadbalancer_test.go
@@ -361,3 +361,30 @@ func TestServiceFlags_String(t *testing.T) {
 		})
 	}
 }
+
+func benchmarkHash(b *testing.B, addr *L3n4Addr) {
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		addr.Hash()
+	}
+}
+
+func BenchmarkL3n4Addr_Hash_IPv4(b *testing.B) {
+	addr := NewL3n4Addr(TCP, net.IPv4(1, 2, 3, 4), 8080, ScopeInternal)
+	benchmarkHash(b, addr)
+}
+
+func BenchmarkL3n4Addr_Hash_IPv6_Short(b *testing.B) {
+	addr := NewL3n4Addr(TCP, net.ParseIP("fd00::1:36c6"), 8080, ScopeInternal)
+	benchmarkHash(b, addr)
+}
+
+func BenchmarkL3n4Addr_Hash_IPv6_Long(b *testing.B) {
+	addr := NewL3n4Addr(TCP, net.ParseIP("2001:0db8:85a3::8a2e:0370:7334"), 8080, ScopeInternal)
+	benchmarkHash(b, addr)
+}
+
+func BenchmarkL3n4Addr_Hash_IPv6_Max(b *testing.B) {
+	addr := NewL3n4Addr(TCP, net.ParseIP("1020:3040:5060:7080:90a0:b0c0:d0e0:f000"), 30303, 100)
+	benchmarkHash(b, addr)
+}