This commit optimizes the L3n4Addr.Hash function. It is used as a key in
Golang maps and therefore needs to be unique, but it does not need to actually
be a secure hash.

The previous implementation was slow issues because it relied on defer
statements, reflection (due to the use of `%+v`), and SHA256. This
function is used during service lookup in Hubble's hot path, therefore
this optimization should reduce the overhead of Hubble:

![image](https://user-images.githubusercontent.com/50564/104613557-a42e3f00-5687-11eb-95b3-317d2e6b101f.png)


Before this commit:

    BenchmarkL3n4Addr_Hash_IPv4-8         	  608378	      1856 ns/op
    BenchmarkL3n4Addr_Hash_IPv6_Short-8   	  620024	      2080 ns/op
    BenchmarkL3n4Addr_Hash_IPv6_Long-8    	  535290	      2168 ns/op

After this commit:

    BenchmarkL3n4Addr_Hash_IPv4-8         	15079537	        81.1 ns/op
    BenchmarkL3n4Addr_Hash_IPv6_Short-8   	 8348496	       149 ns/op
    BenchmarkL3n4Addr_Hash_IPv6_Long-8    	 4463416	       259 ns/op