@gandro 2021-01-14
> Nice! Do you have a flamegraph for "after"?

This is the flamegraph for just the service lookup afterwards. Most time is still spent in Hash, but now it's mostly memory allocation.

![image](https://user-images.githubusercontent.com/50564/104615653-ba3cff00-5689-11eb-99ab-6ae2c822cd02.png)