@gandro 2021-01-14
Pushed an additional benchmark and made the length calculation a bit more tight (24 instead of 32 bytes by default). See also https://github.com/cilium/cilium/pull/14617#discussion_r557568344