@gandro 2021-01-15
Net-Next hit #14598
https://jenkins.cilium.io/job/Cilium-PR-K8s-1.13-net-next/460/
Two errors in the image build that look spurious and unrelated:
https://jenkins.cilium.io/job/Cilium-PR-Ginkgo-Tests-Kernel/4448/
https://jenkins.cilium.io/job/Cilium-PR-K8s-1.20-kernel-4.9/408/