@gandro 2021-01-18
net-next hit #14598 again, as well as an error similar to #12690 - but in a different context: https://jenkins.cilium.io/job/Cilium-PR-K8s-1.13-net-next/468/testReport/junit/Suite-k8s-1/13/K8sServicesTest_Checks_service_across_nodes_Tests_NodePort_BPF_Tests_with_direct_routing_With_host_policy_Tests_NodePort/